# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'ClearGate FX API Documentation'
  description: 'API documentation for the ClearGate FX currency exchange backend.'
  version: 1.0.0
servers:
  -
    url: 'http://exhange.test'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Use the token returned by POST /api/v1/auth/login as a Bearer token.'
security:
  -
    default: []
paths:
...
  /api/v1/auth/change-password:
    put:
      summary: 'Change password'
      operationId: changePassword
      description: "Update the authenticated user's password after verifying the current password."
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'تم تغيير كلمة المرور'
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'تم تغيير كلمة المرور'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: false
                  message: 'كلمة المرور الحالية غير صحيحة'
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: 'كلمة المرور الحالية غير صحيحة'
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: ''
                  example: architecto
                password:
                  type: string
                  description: 'Must be at least 8 characters.'
                  example: ']|{+-0pBNvYg'
              required:
                - current_password
                - password
...